+++ src/store/taskStore.ts
@@ -10,6 +10,7 @@ type state = {
 
 type actions = {
   addTask: (task: taskType) => void;
+  updateTask: (updatedTask: taskType) => void;
   removeTask: (taskId: string) => void;
   toggleComplete: (taskId: string) => void;
   removeCompletedTasks: () => void;
@@ -18,13 +19,20 @@ type actions = {
 const useTaskStore = create<state & actions>()((set) => ({
   // Stores all task items; they are objects of type taskType
   tasks: [],
-
   // Add a task
   addTask: (task: taskType) =>
     set((state) => ({
       tasks: [...state.tasks, task],
     })),
-
+  updateTask: (updatedTask: taskType) =>
+    set((state) => ({
+      tasks: state.tasks.map((task: taskType) => {
+        if (task.id === updatedTask.id) {
+          return updatedTask;
+        }
+        return task;
+      }),
+    })),
   // Remove a task by id
   removeTask: (taskId: string) =>
     set((state) => ({